Added TODOs on how to convert level loader from being hardcoded to something more general.
TODOs are found in the world class. After OS I'll implement it so it's
data driven and so we have more variety and flexibility in creating the
maps. Will also generate a GUI like in project 2 later.

Can someone add a legend detailing all the different types of tiles we
will have so I can  make the GUI?