Fixed some bugs, restructured some code for better gameplay.
+ Fixed bug mentioned in previous commit.
+ Separated the different enemy difficulties so player doesn't die
automatically if a hard enemy spawns.
+ Easy enemies found in the grass level.
+ Medium and hard enemies found in the caves level.
+ See ascii map and map legend for details
+ Do not alter the transitions between grass map to cave map unless you
know what you're doing because there is a serious bug. Current solution
that offsets the transitions avoids the bug entirely so far although
it's not the best solution.

+ More bugs found:
- null exception if you try to select the area of the defeated enemy
- nothing happens when player health is below 0, so player can continue
attacking despite having negative health
- not sure if this is just on my laptop and it's lagging but controls
are not smooth. Like sometimes wouldn't be able to select enemy or even
move the arrow down the menu screen for both battlesys/shop. Not sure if
that's due to the sound being out of sync or XNA being slow at updating
or just my laptop processing too many things though